+++ modules/test/jenkins-results-parser/src/main/java/com/liferay/jenkins/results/parser/testray/TestrayServer.java
@@ -30,7 +30,7 @@ public interface TestrayServer {
 
 	public TestrayCaseType getTestrayCaseType(String testrayCaseTypeName);
 
-	public TestrayProject getTestrayProjectByID(int projectID);
+	public TestrayProject getTestrayProjectByID(long projectID);
 
 	public TestrayProject getTestrayProjectByName(String projectName);
